Move taglink() function to epydoc2stan module
I want to start using it from that module and that was blocked by
an import cycle.
