+++ pydoctor/epydoc2stan.py
@@ -46,6 +46,24 @@ def get_docstring(
     return None, None
 
 
+def taglink(o: model.Documentable, page_url: str, label: Optional[str] = None) -> Tag:
+    if not o.isVisible:
+        o.system.msg("html", "don't link to %s"%o.fullName())
+
+    if label is None:
+        label = o.fullName()
+
+    url = o.url
+    if url.startswith(page_url + '#'):
+        # When linking to an item on the same page, omit the path.
+        # Besides shortening the HTML, this also avoids the page being reloaded
+        # if the query string is non-empty.
+        url = url[len(page_url):]
+
+    ret: Tag = tags.code(tags.a(label, href=url))
+    return ret
+
+
 class _EpydocLinker(DocstringLinker):
 
     def __init__(self, obj: model.Documentable):
