+++ pydoctor/templatewriter/pages/table.py
@@ -1,5 +1,6 @@
 from twisted.web.template import Element, TagLoader, XMLFile, renderer
 
+from pydoctor import epydoc2stan
 from pydoctor.model import Function
 from pydoctor.templatewriter import util
 
@@ -31,7 +32,7 @@ class TableRow(Element):
 
     @renderer
     def name(self, request, tag):
-        return tag.clear()(util.taglink(self.child, self.ob.url, self.child.name))
+        return tag.clear()(epydoc2stan.taglink(self.child, self.ob.url, self.child.name))
 
     @renderer
     def summaryDoc(self, request, tag):
