+++ pydoctor/templatewriter/summary.py
@@ -9,7 +9,7 @@ from twisted.web.template import Element, Tag, TagLoader, XMLFile, renderer, tag
 
 def moduleSummary(modorpack, page_url):
     r = tags.li(
-        util.taglink(modorpack, page_url), ' - ',
+        epydoc2stan.taglink(modorpack, page_url), ' - ',
         epydoc2stan.format_summary(modorpack)
         )
     if modorpack.isPrivate:
@@ -105,7 +105,7 @@ def subclassesFrom(hostsystem, cls, anchors, page_url):
     if name not in anchors:
         r(tags.a(name=name))
         anchors.add(name)
-    r(util.taglink(cls, page_url), ' - ', epydoc2stan.format_summary(cls))
+    r(epydoc2stan.taglink(cls, page_url), ' - ', epydoc2stan.format_summary(cls))
     scs = [sc for sc in cls.subclasses if sc.system is hostsystem and ' ' not in sc.fullName()
            and sc.isVisible]
     if len(scs) > 0:
@@ -185,7 +185,7 @@ class LetterElement(Element):
     @renderer
     def names(self, request, tag):
         def link(obj: model.Documentable) -> Tag:
-            tag = util.taglink(obj, NameIndexPage.filename)
+            tag = epydoc2stan.taglink(obj, NameIndexPage.filename)
             # The "data-type" attribute helps doc2dash figure out what
             # category (class, method, etc.) an object belongs to.
             tag(**{"data-type": obj.kind})
@@ -284,7 +284,7 @@ class IndexPage(Element):
         else:
             root, = self.system.rootobjects
             return tag.clear()(
-                "Start at ", util.taglink(root, self.filename),
+                "Start at ", epydoc2stan.taglink(root, self.filename),
                 ", the root ", root.kind.lower(), ".")
 
     @renderer
@@ -298,7 +298,7 @@ class IndexPage(Element):
     def roots(self, request, tag):
         r = []
         for o in self.system.rootobjects:
-            r.append(tag.clone().fillSlots(root=util.taglink(o, self.filename)))
+            r.append(tag.clone().fillSlots(root=epydoc2stan.taglink(o, self.filename)))
         return r
 
     @renderer
@@ -351,7 +351,7 @@ class UndocumentedSummaryPage(Element):
                           if o.isVisible and not hasdocstring(o)]
         undoccedpublic.sort(key=lambda o:o.fullName())
         for o in undoccedpublic:
-            tag(tags.li(o.kind, " - ", util.taglink(o, self.filename)))
+            tag(tags.li(o.kind, " - ", epydoc2stan.taglink(o, self.filename)))
         return tag
 
 summarypages = [
