+++ pydoctor/templatewriter/util.py
@@ -5,7 +5,6 @@ import os
 
 from pydoctor import model
 from twisted.python.filepath import FilePath
-from twisted.web.template import Tag, tags
 
 
 def srclink(o: model.Documentable) -> Optional[str]:
@@ -18,20 +17,3 @@ def templatefile(filename):
 
 def templatefilepath(filename):
     return FilePath(templatefile(filename))
-
-def taglink(o: model.Documentable, page_url: str, label: Optional[str] = None) -> Tag:
-    if not o.isVisible:
-        o.system.msg("html", "don't link to %s"%o.fullName())
-
-    if label is None:
-        label = o.fullName()
-
-    url = o.url
-    if url.startswith(page_url + '#'):
-        # When linking to an item on the same page, omit the path.
-        # Besides shortening the HTML, this also avoids the page being reloaded
-        # if the query string is non-empty.
-        url = url[len(page_url):]
-
-    ret: Tag = tags.code(tags.a(label, href=url))
-    return ret
